pyproject.toml
[project]
name = "iidx-genre-fixer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytaglib>=3.1.0",
]
